Update instructions for Unit 17 - Activity 5
This commit updates instructions in PT LP 17.1 and FT LP 9.1 by adding
example mongo shell commands for the instructor to reference for
activity 5. Also, changed instructions for students to add 5 documents
instead of 10 given the time constraint and given that the purpose of
the activity is to introduce sorting.

Update README.md in Unit 17, Activity 5 to include instruction to add
documents to animals collection.

Closes #259