Fix issue #352 - unit 18, activity 13
This commit updates activity 13 of unit 18 to address a lack of time
for students to complete the activity. In order to address this,
several snippets of code are provided in the instructions and in
comments found in service-worker.js to reduce the amount of typing
required. The intent of the changes is to draw the student's attention
to the key components of enabling caching for an offline experience
while taking less time to complete the activity.

Additionally, the `Solved` app did not work offline. This commit
fixes the solution so that it works offline.

- Updates README.md with code snippets and additional steps.
- Adds comments and commented out code snippet to `service-worker.js`
- Updates `Solved/public/service-worker.js` to enable offline mode.
- Removes unused paths from static cache in
  `Solved/public/service-worker.js`
- Updates ft lp day 9.4 to reflect the changes made to activity 13.
- Adds missing images to ft lp day 9.4.
- Updates pt lp day 18.2 to reflect the changes made to activity 13.

Fixes #352