Fix issue #370 - Unit 19 - Activity 22
This commit updates unit 19, activity 22 so that the link styling
of the Navbar updates when the location changes.

**Changes to Activity 22:**

Updates `/22-Ins_IntroReactRouter/src/components/NavTabs.js` to
import `useLocation` hook to dynamically set the className props
on the navigation links. Adds a link to the React Router docs in the
comments with a note stating that the Hooks API will be covered in the
following unit.

**Changes to FT LP 10.3 and PT LP 19.3:**

Modifies the lesson plans to reflect these changes. A recommendation
for the instructor to explain what the code does and leave the rest
to be talked about in the subsequent unit on the Hooks APi.

Fixes #370