Fix issue #508 - unit 20 - activity 18
This commit modifies activity `18-Ins_useRef/src/App.js` so that
clicking on the "Add" or "Subtract" buttons after setting teh value
with the "Change" button does not concatenate the numbers as strings.

Updates code snippets in the lesson plans to match the changes in the
activity code. (PT Day 20.3 and FT Day 10.5)

Fixes #508