Fix unit 21 activities
This commit applies several updates and fixes for the activities
in unit 21 and the lesson plans. (p/t day 21.1 and f/t day 11.1)

**Changes to 02-Stu_Recipes activity:**
The Recipe Puppy API has not been working for at least a week. A
connection to a local MongoDB instance and seed data allows students
to continue to get the core concept of the activity without relying
on a third party API. The recipe seed data was sourced by scraping
several Recipe Puppy search result pages.

- Fix manifest in solved/unsolved
- Replace Recipe Puppy API with MongoDB connection and seeds.
- Update instructions to include steps to seed the database.

**Changes to 03-Stu_AJAXBooks activity:**
- Fix unsolved activity failing to compile
- Fix solved activity failing to compile
- Fix error in manifest.json

**Changes to 04-Stu_AJAXFormDelete activity:**
- Add comments to unsolved code to clarify instructions.
- Solved provides initial values for form state in Books.js

**Changes to 05-Stu_ReactRouter activity:**
This activity relies heavily on students using the React Router docs.
The documentation no longer contains examples similar to those in the
solution and use a different approach to routing and receiving route
parameters.

- Upgrade to React Router Dom v5.x
- Update solution to follow the techniques for routing and receiving
  route parameters found in the documentation.
- Fix error in manifest.json
- Call out the fact that the starter code does not work until part 1
  is completed in the instructions.

**Changes to PT LP Day 21.1 and FT LP Day 11.1:**
- Fix broken and missing images
- Fix broken links to activities
- Update code snippets to correspond to changes made in activities.