Fix #449 - Unit 3 - 35-LocationRedirect
This commit updates
01-Class-Content/03-JavaScript/01-Activities/35-LocationRedirect/ so
that users are redirected correctly based on location. The Geolocation
API uses negative numbers for western hemisphere longitudes. All users
west of 100 deg. west are redirected to the "west" page
and all other users are redirected to the "east" page.

Also, "eastcoster" and "westcoaster" terms are changed to "easterner"
and "westerner" respectively. This curriculum is used by many students
who are not located on a coast. :)

Fixes #449