Fix #290 - Unit 9/Activity 10 lesson plan notes
Activity 10 in unit 9 does not behave as described if the activity
directory is a descendent of a folder containing a `package.json` file.

This commit updates the lesson plans with a note for the instructor:

"The following activity, `01-Activities/10-Ins_npm`, will not behave as
described if it is run from within the curriculum repo or another repo.
First, copy the directory to another location so that there is not an
ancestor directory containing a `package.json` file."

Fixes #290