Fix #268 - Unit 14 - Password link in example
This commit removes a bug where the hashed password was leaked when
a user successfully logged in on the `/api/login` route in
`01-Class-Content/14-Full-Stack/04-Supplemental/Sequelize-Passport-Example`.
Updates the route middleware for `/routes/api-routes.js` handling
`POST /api/login` request so that only the user's id and email is
returned in the response.

Fixes #268