Fix #353 and other changes to unit 18/activity 18
This commit updates the instructions, solved and unsolved parts of
activity 18, and the corresponding sections of the part-time and
full-time lesson plans.

The purpose of the update is to remove the sw pre-cache plugin from the
activity and align it with subsequent image gallery activities. The
focus of the activity is to provide students with a first experience of
using a plugin with webpack. The activity requires students to use
documentation and previous activities to refactor the Image Gallery
application so that Webpack generates a minified JS bundle and a
manifest (using the a plugin).

- Update corresponding sections of the p/t and f/t lesson plans for
  changes to activity 18.

- Remove `Solved/Images` and Unsolved/Images` from
  `18-Stu-Webpack-Plugin`. These images are copies of those used in the
  lesson plans and are not required for the activity or activity
  instructions.

- Remove `Solved/public/dist` from activity 18. This should be generated
  at runtime by Webpack.

- Update `service-worker.js` in Solved and Unsolved to align with
  subsequent activities based on the Image Gallery app. App caches
  necessary files so that it functions when offline.

Fixes #353