Fix #481 - Unit 18/Homework starter code
This commit addresses issue #481 and makes a few other fixes to
`01-Class-Content/18-PWA/02-Homework/`.

Adds a fully-functional starter app as described in the instructions.
The starter app is the described app without the ability to function
offline. Students will have to provide a service worker, manifest, and
code to utilize indexedDB to allow data to be entered offline and synced
with the backend when the app is online again. This fixes #481.

Updates the servicer worker in the `Master` app with additional files
to cache to allow the app to function correctly when offline.

Updates the service worker in the `Master` app to cache api requests
so that transaction data is cached for offline viewing.

Fixes #481