Update 18-PWA/01-Activities/23-Stu-Mini-Project
This commit fixes issue #554 in addition to some other fixes
and updates by modifying `18-PWA/01-Activities/23-Stu-Mini-Project`
and the corresponding sections of the PT (day 18.3) and FT (day 9.5)
lesson plans.

**Fix #554**

This fixes the problem reported in issue #554 by updating the path to an
image referenced in `webpack.config` in the `Solved` dir.

Fixes #554

**Fix activity 23 unsolved/server.js**

Fixes a typo causing a "Module not found error" in `Unsolved/server.js`.

**Fix intermittent error in Solved and Unsolved client**

Fixes error thrown trying to generate an article id when `article.title`
is null in `Solved/client/assets/js/utils.js` and
`Unsolved/client/assets/js/topic.js`.

Modifies `createArticleId` to create pseudo-unique ids using
`article.url` with the special characters removed. Occasionally, the
`title` field is null in the data from the news api and should not be
used for generating ids.

Refactors `utils.js` by changing name of `removePunctuationFromString`
to `removeSpecialCharsFromString`.

**Remove sw-precache-webpack-plugin**

The service worker generated by sw-precache-webpack-plugin was not
linked to the front end pages in the solution for activity 23.
Removes the plugin from `webpack.config` and `client/package.json`.

**Update lesson plans for activity 23 changes**

Update the sections corresponding to activity 23 in PT LP Day 18.3 and
FT LP Day 9.5.