Fix #553 - Unit 18 > activity 22 db name
Fixes issue where the seeder script and server were using different
db names in Solved and Unsolved. This commit updates `server.js` in
each case so that the db name matches that used in `seed.js`. This
sticks with the same db name used in previous image gallery activities
found in the unit.

Fixes #533