Fixes unit 18 activity 14 service worker caching
This adds code to fix the caching of the service-worker for activity
14-Stu_Notetaker_PWA of unit 18. The scope of this commit is to fix
the problem of the app not rendering the page when offline.

This commit does not address the lack of offline support for storing
changes to notes when offline. That would require additional code and
probably mean adding more content to the lesson plan to discuss handling
data offline.

Addresses issue #629 in part. Not sure if it warrant closing the issue.