Update unit 18 activity 14 solution caching
Updates the dynamic caching of 14-Stu_Notetaker_PWA/Solved in the
unit 18 content. The condition for using a network first falling back
to cache strategy was checking for "/api" in the url which is not
part of any route in the app. Changes the condition to check for "/all"
or "/find" which are part of the paths for getting all notes or the
details for a single note.