Update unit 9, activity 28
This commit removes requirement to use promises making it a bonus. Also,
reference to asnyc/await has been removed from the instructions for
the bonus.

The motivation for this change is due to the fact that all references
to promises and async/await have been removed from the lesson plans
and activities for teaching these concepts were removed. There is not
enough time in the lesson plan to allow for covering these concepts
on the day of the mini-project.

Fixes #1436