wip: Update m5 act28 html and style
This work in progress commit removes the javascript from the solution
for the module 5 mini-project and updates the styles.

Some elements are included that will be removed from the finished
update and are there as prototypes.

These changes are made as part of an effort to add localStorage to the
mini-project to align it with the challenge for this unit. Because
localStorage is begin added to the mini-proj. requirements, some
features (modal and jQuery UI) are being removed in order to maintain
the current level of difficulty.