Fix mod 05 activities 17-20 - navbar toggler
Fixes bug with the navbar toggler. This bug existed in BS 4 and
persisted after migrating to BS 5. The navbar toggler is only visible
in combination with the navbar-dark or navbar-light classes on the
parent .navbar element to the navbar-toggler-icon. Without, the
navbar-toggler is not visible when the navbar is collapsed for mobile
and tablet screens